Change the sections callback from a function into a class.
This fixes #5, i.e. the issue where sometimes it would crash with an "index out of range" error when generating the debug sections.

Even though it's a very simple fix it took a lot of hours of staring at the libdwarf code to figure out what could go wrong and of fprintf debugging.
The issue was that, sometimes, the call to dbg->de_callback_func inside pro_section.c "failed" (i.e. our script never received it) and returned 0 by default, which means that we never populated debug_sections with the names and that all the sections got an index of 0.
So, even though it was actually generating the content of the sections correctly, they got all concatenated inside one big section (with index 0).

I suspect this issue was caused by the GC collecting our callback by mistake, so instead of a python function, we are now using a class implementing the auto-generated Dwarf_Callback_Func interface to to create an object which should be kept alive by the GC.

To be fair, I'm surprised our `jython -> java -> C library` approach even works so I'm happy, all is well that ends well!