Better error handling.
libdwarf functions are now wrapped inside a python function that does error checking automatically and asserts whenever an error in encountered.
This makes calling those functions easier since we don't have to check for errors each time and we are now sure we are checking for the correct error value.
The function wrapper also automatically appends the Dwarf_Error argument to the call so we don't have to write `err` as the last argument anymore.