Fix line number table corruption by sorting output by address.
The DWARF spec, section 6.2.5, states that "Within a sequence, addresses and
operation pointers may only increase.". Since we are generating only a single
sequence (corresponding to the single compilation unit *_dbg.c), we need to
ensure that all addresses are sorted.

libdwarf silently misbehaves when fed non-monotonic addresses: if the address
goes backwards by just a small amount, it's possible for the computed opcode to
be below the opcode base; libdwarf doesn't check for this and outputs the broken
opcode directly, which will cause the opcode to be interpreted as a random
standard opcode, corrupting the line number table after that point.

Here's a concrete sequence of calls that should trigger the problem:

    dwarf_add_line_entry(dbg, file_index, 0x401fb7L, 1093, 0, True, False)
    dwarf_add_line_entry(dbg, file_index, 0x401fbbL, 1095, 0, True, False)
    dwarf_add_line_entry(dbg, file_index, 0x401fb9L, 1096, 0, True, False)
    dwarf_add_line_entry(dbg, file_index, 0x401fd1L, 1097, 0, True, False)
    dwarf_add_line_entry(dbg, file_index, 0x401fd1L, 1098, 0, True, False)

The third call will add an opcode with value 4, which is interpreted as
DW_LNS_set_file, thereby consuming the next opcode; subsequent entries will be
shifted and will have the wrong file index.