Fix missing instructions at the end of functions.
Ghidra maxAddress is inclusive (it's the last address contained within the
range), but DW_AT_high_pc is exclusive (DWARF v5 standard 2.17.2, "If the value
of the DW_AT_high_pc is of class address, it is the address of the first
location past the last instruction associated with the entity."). Thus, instead
of -1, we actually need a +1 adjustment. This fixes a bug where the last
instruction (e.g. a one-byte "ret") was not being included in function listings
in GDB.